fix more npes
