+++ paper/src/main/java/com/badbones69/crazycrates/paper/cratetypes/QuadCrate.java
@@ -73,7 +73,7 @@ public class QuadCrate implements Listener {
 
                 Crate crate = session.getCrate();
                 Prize prize = crate.pickPrize(player, block.getLocation().add(.5, 1.3, .5));
-                this.crazyManager.givePrize(player, prize, crate);
+                this.plugin.getCrazyHandler().givePrize(player, prize, crate);
 
                 ItemBuilder itemBuilder = ItemBuilder.convertItemStack(prize.getDisplayItem());
                 itemBuilder.addLore(new Random().nextInt(Integer.MAX_VALUE) + ""); // Makes sure items don't merge
