+++ paper/src/main/java/com/badbones69/crazycrates/paper/cratetypes/QuickCrate.java
@@ -57,7 +57,7 @@ public class QuickCrate implements Listener {
                 if (keysUsed >= crate.getMaxMassOpen()) break;
 
                 Prize prize = crate.pickPrize(player);
-                crazyManager.givePrize(player, prize, crate);
+                plugin.getCrazyHandler().givePrize(player, prize, crate);
                 plugin.getServer().getPluginManager().callEvent(new PlayerPrizeEvent(player, crate, crate.getName(), prize));
 
                 if (prize.useFireworks()) methods.firework(loc.clone().add(.5, 1, .5));
@@ -82,7 +82,7 @@ public class QuickCrate implements Listener {
             }
 
             Prize prize = crate.pickPrize(player, loc.clone().add(.5, 1.3, .5));
-            crazyManager.givePrize(player, prize, crate);
+            plugin.getCrazyHandler().givePrize(player, prize, crate);
             plugin.getServer().getPluginManager().callEvent(new PlayerPrizeEvent(player, crate, crate.getName(), prize));
             ItemStack displayItem = prize.getDisplayItem();
             NBTItem nbtItem = new NBTItem(displayItem);
