+++ paper/src/main/java/com/badbones69/crazycrates/paper/cratetypes/Roulette.java
@@ -90,7 +90,7 @@ public class Roulette implements Listener {
                         crazyManager.endCrate(player);
                         Prize prize = crate.getPrize(inv.getItem(13));
 
-                        methods.checkPrize(prize, crazyManager, plugin, player, crate);
+                        plugin.getCrazyHandler().checkPrize(prize, player, crate);
 
                         crazyManager.removePlayerFromOpeningList(player);
